feat: 抽出 chat 和 setup 组件
