+++ spring-ai-alibaba-studio/ui/src/components/Chat/index.tsx
@@ -15,16 +15,22 @@
  */
 
 import { useEffect, useState, useRef, memo } from 'react';
-import { Flex, Card, Button, Checkbox, Input, Spin, Image, Divider } from 'antd';
-import Setup from '../Setup';
-import { useParams } from 'ice';
+import {
+  Flex,
+  Card,
+  Button,
+  Checkbox,
+  Input,
+  Spin,
+  Image,
+  Divider,
+} from 'antd';
 import {
   ChatModelData,
   ChatModelResultData,
   ModelType,
 } from '@/types/chat_model';
 import chatModelsService from '@/services/chat_models';
-import { RightPanelValues } from '../types';
 import { RobotOutlined, UserOutlined } from '@ant-design/icons';
 import styles from './index.module.css';
 import { ChatOptions, ImageOptions } from '@/types/options';
@@ -32,69 +38,12 @@ import { ChatOptions, ImageOptions } from '@/types/options';
 type Props = {
   modelData: ChatModelData;
   modelType: ModelType;
-};
-
-type Params = {
-  model_name: string;
+  modelOptions: ChatOptions | ImageOptions | undefined;
+  prompt: string;
 };
 
 const ChatModel = memo((props: Props) => {
-  const { modelData, modelType } = props;
-  // 路径参数
-  const params = useParams<Params>();
-
-  const [initialValues, setInitialValues] = useState<RightPanelValues>({
-    initialChatConfig: {
-      model: 'qwen-plus',
-      temperature: 0.85,
-      top_p: 0.8,
-      seed: 1,
-      enable_search: false,
-      top_k: 0,
-      stop: [],
-      incremental_output: false,
-      repetition_penalty: 1.1,
-      tools: [],
-    },
-    initialImgConfig: {
-      model: 'wanx-v1',
-      responseFormat: '',
-      n: 1,
-      size: '1024*1024',
-      style: '<auto>',
-      seed: 0,
-      ref_img: '',
-      ref_strength: 0,
-      ref_mode: '',
-      negative_prompt: '',
-    },
-    initialTool: {},
-  });
-  const [modelOptions, setModelOptions] = useState<
-    ChatOptions | ImageOptions
-  >();
-  const [prompt, setPrompt] = useState('');
-
-  // 当 modelData.chatOptions 发生变化时同步更新 initialValues
-  useEffect(() => {
-    if (params.model_name != modelData.name) {
-      return;
-    }
-    // 该属性不能传
-    if (modelData != null && modelData.chatOptions != null) {
-      delete modelData.chatOptions.proxyToolCalls;
-    }
-    setInitialValues((prev) => ({
-      initialChatConfig: { ...modelData.chatOptions },
-      initialImgConfig: { ...modelData.imageOptions },
-      initialTool: {},
-    }));
-    if (modelData.modelType == ModelType.CHAT) {
-      setModelOptions(modelData.chatOptions);
-    } else if (modelData.modelType == ModelType.IMAGE) {
-      setModelOptions(modelData.imageOptions);
-    }
-  }, [modelData]);
+  const { modelData, modelType, modelOptions, prompt } = props;
 
   const [inputValue, setInputValue] = useState('');
   const [isStream, setIsStream] = useState(false);
@@ -110,16 +59,12 @@ const ChatModel = memo((props: Props) => {
   };
 
   const [messages, setMessages] = useState(
-    [] as Array<{ type: string; content: JSX.Element | string }>,
+    [] as Array<{
+      type: string;
+      content: JSX.Element | string;
+      isClear?: boolean;
+    }>,
   );
-
-  const handleOptions = (options: ChatOptions | ImageOptions) => {
-    setModelOptions(options);
-  };
-
-  const handlePrompt = (prompt: string) => {
-    setPrompt(prompt);
-  };
 
   const runModel = async () => {
     try {
@@ -209,12 +154,12 @@ const ChatModel = memo((props: Props) => {
   }, [messages]);
 
   return (
-    <Flex justify="space-between" style={{ height: '100%' }}>
-      <Flex vertical style={{ marginRight: 20, flexGrow: 1, height: '100%' }}>
-        <div className={styles['message-wrapper']}>
-          {messages.map((message: any, index) => {
-            return (
-              <><Flex
+    <Flex vertical style={{ marginRight: 20, flexGrow: 1, height: '100%' }}>
+      <div className={styles['message-wrapper']}>
+        {messages.map((message: any, index) => {
+          return (
+            <>
+              <Flex
                 key={index}
                 className={styles['message']}
                 style={{
@@ -247,37 +192,31 @@ const ChatModel = memo((props: Props) => {
                   <UserOutlined className={styles['message-icon']} />
                 )}
               </Flex>
-                {message.isClear && <Divider>上下文已结束</Divider>}</>
-            );
-          })}
-        </div>
+              {message.isClear && <Divider>上下文已结束</Divider>}
+            </>
+          );
+        })}
+      </div>
 
-        <Flex vertical>
-          <TextArea
-            autoSize={{ minRows: 3 }}
-            style={{ marginBottom: 20 }}
-            value={inputValue}
-            onChange={handleInputChange}
-          />
-          <Flex style={{ flexDirection: 'row-reverse' }}>
-            <Flex style={{ width: 300 }} align="center" justify="space-around">
-              <Button onClick={cleanHistory}>清空</Button>
-              <Checkbox checked={isStream} onChange={handleStreamChange}>
-                流式响应
-              </Checkbox>
-              <Button onClick={runModel} disabled={disabled}>
-                运行
-              </Button>
-            </Flex>
+      <Flex vertical>
+        <TextArea
+          autoSize={{ minRows: 3 }}
+          style={{ marginBottom: 20 }}
+          value={inputValue}
+          onChange={handleInputChange}
+        />
+        <Flex style={{ flexDirection: 'row-reverse' }}>
+          <Flex style={{ width: 300 }} align="center" justify="space-around">
+            <Button onClick={cleanHistory}>清空</Button>
+            <Checkbox checked={isStream} onChange={handleStreamChange}>
+              流式响应
+            </Checkbox>
+            <Button onClick={runModel} disabled={disabled}>
+              运行
+            </Button>
           </Flex>
         </Flex>
       </Flex>
-      <Setup
-        modelType={modelData.modelType}
-        initialValues={initialValues}
-        onChangeConfig={handleOptions}
-        onChangePrompt={handlePrompt}
-      />
     </Flex>
   );
 });
