--- spring-ai-alibaba-studio/ui/src/pages/run/clients/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Empty } from 'antd';
-
-function Clients() {
-  return (
-    <div style={{ padding: 20, height: '100%', display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-      <Empty />
-    </div>
-  );
-}
-
-export default Clients;
